Initial Commit
This creates a new start for Endless Sky, wherein the player starts in the same place as vanilla, with the same bank conversation. The key difference is that the player has an old Lampyrid Mk II Transport they've been fixing up for years.

Key difference from the person Lampyrid: has two turret hardpoints, which at the start have 1 blaster turret and 1 anti missile turret.

This is mainly due to the fact that otherwise the player will attract pirates right from the very first launch.

I am considering dropping that to a single turret hardpoint (more in keeping with the show) but bump the engines up a bit so that the player has a better shot at running away. Because if they only have one turret (probably an AM), they are going to attract pirates regularly.